requirements.txt
streamlit
pypdf
streamlit-drawable-canvas
Pillow
reportlab
pymupdf
numpy
